requirements.txt
bcrypt==4.0.1
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
flake8==6.0.0
future==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
netmiko==4.1.2
ntc-templates==3.1.0
paramiko==2.12.0
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==0.21.0
PyYAML==6.0
scp==0.14.4
six==1.16.0
tenacity==8.1.0
textfsm==1.1.2
tomli==2.0.1
